change solutions report btn
