+++ app/layout/project/sidebar/scenario/solutions/overview/index.tsx
@@ -9,11 +9,7 @@ import { getScenarioEditSlice } from 'store/slices/scenarios/edit';
 import { motion } from 'framer-motion';
 
 import { useProject } from 'hooks/projects';
-import {
-  useScenario,
-  useDownloadScenarioReport,
-  useDownloadSolutionsSummary,
-} from 'hooks/scenarios';
+import { useScenario, useDownloadScenarioReport } from 'hooks/scenarios';
 import { useSolution, useBestSolution } from 'hooks/solutions';
 import { useToasts } from 'hooks/toast';
 
@@ -34,7 +30,6 @@ export const SolutionsOverview = (): JSX.Element => {
   const { pid, sid } = query as { pid: string; sid: string };
 
   const [PDFLoader, setPDFLoader] = useState<boolean>(false);
-  const [solutionsReportLoader, setSolutionsReportLoader] = useState<boolean>(false);
   const [showTable, setShowTable] = useState<boolean>(false);
   const { addToast } = useToasts();
 
@@ -71,8 +66,6 @@ export const SolutionsOverview = (): JSX.Element => {
     runId: `${(selectedSolutionData || bestSolutionData)?.runId}`,
   });
 
-  const downloadSolutionsSummary = useDownloadSolutionsSummary();
-
   const SOLUTION_DATA = selectedSolutionData || bestSolutionData;
 
   const isBestSolution =
@@ -82,31 +75,6 @@ export const SolutionsOverview = (): JSX.Element => {
   const solutionIsLoading =
     (bestSolutionisFetching && !bestSolutionisFetched) ||
     (selectedSolutionisFetching && !selectedSolutionisFetched);
-
-  const onDownloadSolutionsSummary = useCallback(() => {
-    setSolutionsReportLoader(true);
-
-    downloadSolutionsSummary.mutate(
-      { id: pid },
-      {
-        onError: () => {
-          addToast(
-            'download-error',
-            <>
-              <h2 className="font-medium">Error!</h2>
-              <ul className="text-sm">Solutions report not downloaded</ul>
-            </>,
-            {
-              level: 'error',
-            }
-          );
-        },
-        onSettled: () => {
-          setSolutionsReportLoader(false);
-        },
-      }
-    );
-  }, [downloadSolutionsSummary, addToast, pid]);
 
   const onDownloadReport = useCallback(() => {
     setPDFLoader(true);
@@ -207,21 +175,6 @@ export const SolutionsOverview = (): JSX.Element => {
                   settings={layerSettings.solution}
                 />
               </div>
-
-              <Button
-                theme="primary-alt"
-                size="base"
-                className="mb-4 flex h-12 overflow-hidden uppercase"
-                disabled={solutionsReportLoader}
-                onClick={onDownloadSolutionsSummary}
-              >
-                <Loading
-                  visible={solutionsReportLoader}
-                  className="absolute bottom-0 left-0 right-0 top-0 z-40 flex h-full w-full items-center justify-center bg-gray-800 bg-opacity-90"
-                  iconClassName="w-10 h-10 text-primary-500"
-                />
-                Export solutions summary
-              </Button>
 
               <Button
                 theme="primary-alt"
